HIP: Add input and output wrappers for faster load and store address computation
On HIP if all addresses from a buffer are smaller than 32 bit, then addressing calculations can be done
in 32 bits saving a few instructions and registers each time a buffer is accessed.
Add a new option useStrict32BitAddress for this optimization. 1 means enable, -1 disable, 0 to infer from size.
Note that this can only be done if the addresses (in bytes) are < 32 bit not the indices (number of elements)
therefore this is not the same as useUint64, but it has no effect when useUint64 is 1.

Co-authored-by: Anton Gorenko <anton@streamhpc.com>